+++ jest.config.js
@@ -1,6 +1,5 @@
 module.exports = {
   preset: 'ts-jest/presets/js-with-ts',
-  testEnvironment: 'jsdom',
   coverageReporters: ['lcov'],
   coverageDirectory: './coverage',
   roots: ['<rootDir>/test'],