+++ src/public/mathlive.ts
@@ -17,41 +17,18 @@
  */
 
 import { RemoteVirtualKeyboard } from '../editor-mathfield/remote-virtual-keyboard';
-import {
-  RemoteVirtualKeyboardOptions,
-  TextToSpeechOptions,
-  AutoRenderOptions,
-} from './options';
+import { RemoteVirtualKeyboardOptions, AutoRenderOptions } from './options';
 
 export * from './commands';
 export * from './core';
 export * from './options';
 export * from './mathfield';
 export * from './mathfield-element';
+export * from './mathlive-ssr';
 
 export declare function makeSharedVirtualKeyboard(
   options?: Partial<RemoteVirtualKeyboardOptions>
 ): RemoteVirtualKeyboard;
-
-export declare function convertLatexToMarkup(
-  text: string,
-  options?: {
-    mathstyle?: 'displaystyle' | 'textstyle';
-    format?: string;
-  }
-): string;
-
-export declare function convertLatexToMathMl(
-  latex: string,
-  options?: Partial<{
-    generateID: boolean;
-  }>
-): string;
-
-export declare function convertLatexToSpeakableText(
-  latex: string,
-  options?: Partial<TextToSpeechOptions>
-): string;
 
 export declare function renderMathInDocument(options?: AutoRenderOptions): void;